Improved keyexchange_remote and bug fixes to previous shellcheck updates.
Improvements to keyexchange_remote, it is now broken into separate preflight and command. Added a check that the key type is an expected value.

Also contains a fix to recent shellcheck commits, "${var:?[@]}" passes shellcheck but does not work as expected in Bash, in a for loop only the first element of the array is iterated.